requirements.txt
alembic==0.9.1
arrow==0.10.0
appdirs==1.4.3
cfenv==0.5.3
click==6.7
connexion==1.1.11
crochet==1.7.0
ecdsa==0.13
flake8==2.5.4
Flask==0.12
Flask-Cors==3.0.3
Flask-Migrate==1.8.0
Flask-Moment==0.5.1
Flask-OAuth==0.12
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Twisted==0.1.2
Flask-WTF==0.14.2
future==0.16.0
gunicorn==19.7.1
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
oauth2==1.9.0.post1
oauthlib==2.0.2
packaging==16.8
phonenumbers==8.4.1
psycopg2==2.7.1
pycrypto==2.6.1
pylint==1.6.4
pyparsing==2.2.0
pystache==0.5.4
pytest-cov==2.4.0
python-dateutil==2.6.0
python-editor==1.0.3
python-jose==1.3.2
python-openid==2.2.5
redis==2.10.6
requests==2.11.1
requests-oauthlib==0.8.0
six==1.10.0
structlog==17.2.0
SQLAlchemy==1.1.9
testing.postgresql==1.3.0
treq==17.3.1
Werkzeug==0.12.1
WTForms==2.1
